requirements.txt
python-telegram-bot[job-queue]==21.3
aiosqlite==0.20.0
pytz==2024.1
